# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: SaaS Onboarding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whatsapp/webhook:
    get:
      operationId: verifyWhatsappWebhook
      tags: [whatsapp]
      summary: Verify WhatsApp webhook (Meta challenge)
      parameters:
        - name: hub.mode
          in: query
          schema:
            type: string
        - name: hub.challenge
          in: query
          schema:
            type: string
        - name: hub.verify_token
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Challenge echoed
        "403":
          description: Forbidden
    post:
      operationId: receiveWhatsappWebhook
      tags: [whatsapp]
      summary: Receive WhatsApp webhook events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappWebhookPayload"
      responses:
        "200":
          description: Accepted
  /whatsapp/send:
    post:
      operationId: sendWhatsappMessage
      tags: [whatsapp]
      summary: Send a WhatsApp message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappSendInput"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappSendResult"
  /whatsapp/sessions:
    get:
      operationId: listWhatsappSessions
      tags: [whatsapp]
      summary: List WhatsApp sessions
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhatsappSession"
...
components:
  schemas:
...
    WhatsappSession:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: ["string", "null"]
        phone:
          type: string
        status:
          type: string
          enum: [active, completed, abandoned]
        flowStep:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - phone
        - status
        - flowStep
        - createdAt
        - updatedAt
    WhatsappWebhookPayload:
      type: object
      properties:
        object:
          type: string
        entry:
          type: array
          items:
            type: object
      required:
        - object
    WhatsappSendInput:
      type: object
      properties:
        to:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        tenantId:
          type: string
      required:
        - to
        - message
    WhatsappSendResult:
      type: object
      properties:
        messageId:
          type: ["string", "null"]
        status:
          type: string
      required:
        - status
...